Implementa el slider doble usando la librería materials de google

+++ app/src/main/java/com/example/pruebas/InvoiceResponse.java
@@ -12,8 +12,8 @@ public class InvoiceResponse {
     public int getNumFacturas() { return numFacturas; }
     public List<Invoice> getFacturas() { return facturas; }
 
-    public double getMaxImporte() {
-        double maxImporte = Double.MIN_VALUE; // Inicializamos con el valor más bajo posible
+    public float getMaxImporte() {
+        float maxImporte = Float.MIN_VALUE; // Inicializamos con el valor más bajo posible
 
         // Verificamos si la lista de facturas no es null ni vacía
         if (facturas != null && !facturas.isEmpty()) {
